修复了数据过大溢出的错误，增加了calloc的检验

+++ CA_dll/CA.h
@@ -7,11 +7,11 @@
 #define true  0x01
 #define false 0x00
 
-int factorial(int i,int end);
-int arrangementNumber(int n,int m);
-int combinationNumber(int n,int m);
-int** arrangementArray(int* array,int length,int select,int *total);
-int** combinationArray(int* array,int length,int select,int *total);
+long long factorial(int i,int end);
+long long arrangementNumber(int n,int m);
+long long combinationNumber(int n,int m);
+int** arrangementArray(int* array,int length,int select,long long *total);
+int** combinationArray(int* array,int length,int select,long long *total);
 void print(int** array,int total,int size);
 
 #endif
